Some slightly better built artifact handling and template checking. TODO: More manual tests and unit tests
