+++ package.nls.json
@@ -195,13 +195,15 @@
     "AWS.lambda.import.importError": "Error importing Lambda function {0}: {1}",
     "AWS.lambda.import.fileNotFound": "Handler file {0} not found in imported function.",
     "AWS.lambda.import.noWorkspaceFolders": "Open a workspace before importing a Lambda function.",
+    "AWS.lambda.upload.title": "Select Upload Type",
     "AWS.lambda.upload.prebuiltZip": "Built function in a ZIP archive",
-    "AWS.lambda.upload.prebuiltZip.detail": "The Toolkit will upload the selected ZIP archive.",
+    "AWS.lambda.upload.prebuiltZip.detail": "AWS Toolkit will upload the selected ZIP archive.",
     "AWS.lambda.upload.prebuiltDir": "Built function in a directory",
-    "AWS.lambda.upload.prebuiltDir.detail": "The Toolkit will upload a ZIP of the selected directory.",
+    "AWS.lambda.upload.prebuiltDir.detail": "AWS Toolkit will upload a ZIP of the selected directory.",
     "AWS.lambda.upload.unbuiltDir": "Unbuilt function in a directory",
-    "AWS.lambda.upload.unbuiltDir.detail": "The Toolkit will attempt to build the selected directory using the sam build command.",
-    "AWS.lambda.upload.confirm": "This will immediately publish the selected code as a new version of Lambda: {0}.\n\nThe Toolkit cannot guarantee that the built code will work.\n\nContinue?",
+    "AWS.lambda.upload.unbuiltDir.detail": "AWS Toolkit will attempt to build the selected directory using the sam build command.",
+    "AWS.lambda.upload.confirm": "This will immediately publish the selected code as a new version of Lambda: {0}.\n\nAWS Toolkit cannot guarantee that the built code will work.\n\nContinue?",
+    "AWS.lambda.upload.handlerNotFound": "AWS Toolkit can't find a file corresponding to handler: {0} at filepath {1}.\n\nThis directory likely will not work with this function.\n\nProceed with upload anyway?",
     "AWS.log.fileLocation": "Error logs for this session are permanently stored in {0}",
     "AWS.log.invalidLevel": "Invalid log level: {0}",
     "AWS.message.loading": "Loading...",
